refactor: API 로직 모듈 분리
- 효율적인 유지보수를 위해 기존 소리터 서비스에 종속되어 있던 API 로직을 packages/api 모듈로 분리했습니다.
