--- apps/soritor/src/hooks/mutations/useMutationDeleteUser.ts
@@ -1,14 +1,0 @@
-import { useMutation } from "@tanstack/react-query";
-
-import { deleteUserInfo } from "@/api/user";
-
-export const useMutationDeleteUser = () => {
-  const mutation = useMutation({
-    mutationFn: deleteUserInfo,
-    onMutate: () => {
-      return { mutationKey: "deleteUser" };
-    },
-  });
-
-  return mutation;
-};
