--- apps/soritor/src/hooks/mutations/useMutationSubmitSurvey.ts
@@ -1,13 +1,0 @@
-import { useMutation } from "@tanstack/react-query";
-
-import { submitSurvey } from "@/api/survey";
-
-import { SurveyFormSchemaType } from "@/hooks/useSurveyForm";
-
-export const useMutationSubmitSurvey = () => {
-  const mutation = useMutation({
-    mutationFn: (data: SurveyFormSchemaType) => submitSurvey(data),
-  });
-
-  return mutation;
-};
