+++ packages/api/.eslintrc.cjs
@@ -1,0 +1,4 @@
+module.exports = {
+  root: true,
+  extends: ["@uket/eslint-config/react-internal.js"],
+};
