+++ packages/api/package.json
@@ -1,0 +1,43 @@
+{
+  "name": "@uket/api",
+  "version": "1.0.0",
+  "author": "",
+  "license": "ISC",
+  "description": "",
+  "main": "dist/index.js",
+  "type": "module",
+  "types": "dist/index.d.ts",
+  "scripts": {
+    "lint": "eslint . --max-warnings 0",
+    "build": "vite build"
+  },
+  "dependencies": {
+    "@lukemorales/query-key-factory": "^1.3.4",
+    "@uket/util": "workspace:*",
+    "@uket/ui": "workspace:*",
+    "@tanstack/react-query": "^5.29.0",
+    "@tanstack/react-query-devtools": "^5.29.0",
+    "axios": "^1.6.8"
+  },
+  "devDependencies": {
+    "@types/node": "^20.12.5",
+    "@types/react": "^18.3.2",
+    "@types/react-dom": "^18.3.0",
+    "@uket/eslint-config": "workspace:*",
+    "@uket/typescript-config": "workspace:*",
+    "@vitejs/plugin-react": "^4.2.1",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
+    "typescript": "^5.7.3",
+    "vite": "^5.2.11",
+    "vite-plugin-dts": "^4.5.0"
+  },
+  "exports": {
+    "./error/*": "./src/error/*.ts",
+    "./mutations/*": "./src/mutations/*.ts",
+    "./queries/*": "./src/queries/*.ts",
+    "./types/*": "./src/types/*.ts",
+    "./instance": "./src/instance.ts",
+    ".": "./src/index.ts"
+  }
+}
