+++ packages/api/src/index.ts
@@ -1,0 +1,9 @@
+export { ReactQueryDevtools } from "@tanstack/react-query-devtools";
+export {
+  QueryCache,
+  QueryClient,
+  QueryClientProvider,
+  useQueryClient,
+} from "@tanstack/react-query";
+
+export type * as DefaultErrorType from "./types/index.d.ts";
