+++ packages/api/src/queries/ticket.ts
@@ -1,32 +1,33 @@
 import { useQuery } from "@tanstack/react-query";
 import { createQueryKeys } from "@lukemorales/query-key-factory";
 
-import { getTicketQRCode } from "@/api/ticket";
-
-import { TicketItem } from "@/types/ticketType";
-
-import { createBlobURL } from "@/utils/handleTicket";
+import { MyTicketQRCodeResponse, TicketItem } from "../types/ticket";
+import { fetcher } from "../instance";
 
 export const ticket = createQueryKeys("ticket", {
   qrcode: (ticketId: TicketItem["ticketId"]) => ({
     queryKey: ["qrcode", ticketId],
-    queryFn: () => getTicketQRCode(ticketId),
+    queryFn: async () => {
+      const { data } = await fetcher.get<MyTicketQRCodeResponse>(
+        `/tickets/${ticketId}/qrcode`,
+        {
+          mode: "BOUNDARY",
+          responseType: "blob",
+        },
+      );
+
+      return data;
+    },
   }),
 });
 
-/**
- * 공연 티켓의 QR 코드를 발급받습니다.
- * @param {TicketItem["ticketId"]} ticketId
- * @param {TicketItem["ticketStatus"]} ticketStatus
- * @returns {any}
- */
 export const useQueryTicketQrcode = (
   ticketId: TicketItem["ticketId"],
   ticketStatus: TicketItem["ticketStatus"],
 ) => {
   return useQuery({
     ...ticket.qrcode(ticketId),
-    select: data => createBlobURL(data),
+    select: data => URL.createObjectURL(data as any),
     staleTime: 0,
     enabled: !!ticketId && ticketStatus !== "입금 확인중",
   });
