+++ packages/api/tsconfig.json
@@ -1,0 +1,11 @@
+{
+  "extends": "@uket/typescript-config/vite.json",
+  // "compilerOptions": {
+  //   "baseUrl": ".",
+  //   "paths": {
+  //     "@api/*": ["./src/*"]
+  //   }
+  // },
+  "include": ["src", "vite.config.ts", "src/vite-env.d.ts"],
+  "exclude": ["node_modules"]
+}
