+++ packages/api/vite.config.ts
@@ -1,0 +1,32 @@
+import dts from "vite-plugin-dts";
+import { defineConfig } from "vite";
+import react from "@vitejs/plugin-react";
+
+export default defineConfig({
+  plugins: [
+    react(),
+    dts({
+      tsconfigPath: "./tsconfig.json",
+      outDir: "dist",
+      entryRoot: "src",
+      exclude: ["vite.config.ts"],
+    }),
+  ],
+  build: {
+    lib: {
+      entry: "src/index.ts",
+      name: "@uket/api",
+      formats: ["es"],
+      fileName: () => "index.js",
+    },
+    rollupOptions: {
+      external: ["react", "react-dom"],
+      output: {
+        globals: {
+          react: "React",
+          "react-dom": "ReactDOM",
+        },
+      },
+    },
+  },
+});
